fix(engine): harden SDL3 input semantics and event flow
Impact:
- Restores reliable SDL3 input behavior across typing, selection, camera drag, and wheel interaction.
- Prevents HiDPI viewport/backbuffer mismatches by tracking pixel-size changes.
- Documents the current IME composition caveat for follow-up work.

What:
- Added window-scoped text-input control and engine bindings to enable/disable text input on focused input lifecycle.
- Updated SDL event handling for SDL3:
  - handle `SDL_EVENT_WINDOW_PIXEL_SIZE_CHANGED`
  - normalize mouse wheel for `SDL_MOUSEWHEEL_FLIPPED`
  - emit `text-editing` from `SDL_EVENT_TEXT_EDITING`
  - keep gamepad input-state timestamps in milliseconds.
- Migrated mouse state storage to float position/delta/wheel values.
- Updated Lua state routing and input model to accept `text-editing` events.
- Fixed Lua mouse-button consumers to use boolean `payload.state` semantics.
- Added SDL3 migration note under docs/dev/notes.

Why:
- SDL3 changed text-input activation, mouse precision/event semantics, and window pixel-size behavior.
- Remaining SDL2 assumptions caused real regressions after the migration.

How:
- Refactored `WindowSdl` and `Engine` input/window pathways.
- Extended `engine-events` + `state-base` + input routing for `text-editing`.
- Updated affected Lua input/selection paths and tests.

Risk:
- Input plumbing is cross-cutting; regressions could appear in niche widgets.
- Mitigated by full unit/integration/e2e suite runs and no legacy fallback shims.

Testing:
- `make build` (passed)
- `SKIP_KEYRING_TESTS=1 XDG_DATA_HOME=/tmp/space/tests/xdg-data SPACE_DISABLE_AUDIO=1 SPACE_ASSETS_PATH=$(pwd)/assets make test` (passed)
- `SPACE_DISABLE_AUDIO=1 SPACE_ASSETS_PATH=$(pwd)/assets make test-e2e` (passed)

Follow-ups:
- Implement full IME composition rendering in input widgets using `text-editing` payload (`text`, `start`, `length`).

References:
- Refs: SDL3 migration guide (`/tmp/space/sdl-migration-guide.md`)

+++ src/input_mouse_state.cpp
@@ -5,10 +5,10 @@
 void MouseState::begin_frame()
 {
     std::copy(std::begin(currentButtons), std::end(currentButtons), std::begin(previousButtons));
-    xrel = 0;
-    yrel = 0;
-    wheelX = 0;
-    wheelY = 0;
+    xrel = 0.0F;
+    yrel = 0.0F;
+    wheelX = 0.0F;
+    wheelY = 0.0F;
 }
 
 void MouseState::update_from_mask(SDL_MouseButtonFlags mask)
@@ -19,7 +19,7 @@ void MouseState::update_from_mask(SDL_MouseButtonFlags mask)
     }
 }
 
-void MouseState::set_motion(int newX, int newY, int deltaX, int deltaY)
+void MouseState::set_motion(float newX, float newY, float deltaX, float deltaY)
 {
     x = newX;
     y = newY;
@@ -27,7 +27,7 @@ void MouseState::set_motion(int newX, int newY, int deltaX, int deltaY)
     yrel += deltaY;
 }
 
-void MouseState::add_wheel(int deltaX, int deltaY)
+void MouseState::add_wheel(float deltaX, float deltaY)
 {
     wheelX += deltaX;
     wheelY += deltaY;
